+++ src/routes/+layout.svelte
@@ -4,7 +4,7 @@
   import TabBar from '$lib/components/TabBar.svelte';
   import Toast from '$lib/components/Toast.svelte';
   import LoginForm from '$lib/components/LoginForm.svelte';
-  import { user } from '$lib/stores';
+  import { user, initDark } from '$lib/stores';
   import { onAuth } from '$lib/firebase-client';
   import { onMount } from 'svelte';
 
@@ -13,6 +13,7 @@
   let authed = $state(false);
 
   onMount(() => {
+    initDark();
     return onAuth((u) => {
       user.set(u);
       authed = true;